overwrite default react files with old app files
