+++ react-ui/src/index.js
@@ -4,5 +4,9 @@ import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(
+  <App
+    apiURL={ "http://localhost:3001/api" }
+  />, document.getElementById('root')
+);
 registerServiceWorker();
