+++ react-ui/src/style/project_card.css
