+++ react-ui/src/style/project_pop_up.css
